requirements.txt
torch            # Of de nieuwste versie compatibel met je hardware
torchvision      # Compatibel met de versie van torch
transformers     # Hugging Face Transformers-bibliotheek voor GPT-2
Pillow           # Voor het laden en verwerken van afbeeldingen